requirements.txt
faiss-cpu
fastapi
uvicorn
fastembed
numpy
python-dotenv
requests